pyproject.toml
[tool.poetry]
name = "gaboon_test"
version = "0.1.0"
description = ""
authors = ["Marco Peyfuss <@mpeyfuss>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
gaboon-cli = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
